# Speed up ConfigRawParams and ConfigRawParamsTree reload time