--- a/GCSViews/ConfigurationView/ConfigPlanner.cs
+++ b/GCSViews/ConfigurationView/ConfigPlanner.cs
@@ -133,6 +133,7 @@ namespace MissionPlanner.GCSViews.ConfigurationView
             SetCheckboxFromConfig("autoParamCommit", CHK_AutoParamCommit);
             SetCheckboxFromConfig("ShowNoFly", chk_shownofly);
             SetCheckboxFromConfig("Params_BG", CHK_params_bg);
+            SetCheckboxFromConfig("SlowMachine", chk_slowMachine);
 
             // this can't fail because it set at startup
             NUM_tracklength.Value = Settings.Instance.GetInt32("NUM_tracklength", 200);
@@ -985,6 +986,11 @@ namespace MissionPlanner.GCSViews.ConfigurationView
         private void CHK_params_bg_CheckedChanged(object sender, EventArgs e)
         {
             Settings.Instance["Params_BG"] = CHK_params_bg.Checked.ToString();
+        }
+
+        private void chk_slowMachine_CheckedChanged(object sender, EventArgs e)
+        {
+            Settings.Instance["SlowMachine"] = chk_slowMachine.Checked.ToString();
         }
     }
 }