--- a/GCSViews/ConfigurationView/ConfigRawParamsTree.Designer.cs
+++ b/GCSViews/ConfigurationView/ConfigRawParamsTree.Designer.cs
@@ -50,6 +50,7 @@
             this.label2 = new System.Windows.Forms.Label();
             this.BUT_commitToFlash = new MissionPlanner.Controls.MyButton();
             this.chk_modified = new System.Windows.Forms.CheckBox();
+            this.BUT_refreshTable = new MissionPlanner.Controls.MyButton();
             ((System.ComponentModel.ISupportInitialize)(this.Params)).BeginInit();
             this.SuspendLayout();
             // 
@@ -215,8 +216,16 @@
             this.chk_modified.UseVisualStyleBackColor = true;
             this.chk_modified.CheckedChanged += new System.EventHandler(this.chk_modified_CheckedChanged);
             // 
+            // BUT_refreshTable
+            // 
+            resources.ApplyResources(this.BUT_refreshTable, "BUT_refreshTable");
+            this.BUT_refreshTable.Name = "BUT_refreshTable";
+            this.BUT_refreshTable.UseVisualStyleBackColor = true;
+            this.BUT_refreshTable.Click += new System.EventHandler(this.BUT_refreshTable_Click);
+            // 
             // ConfigRawParamsTree
             // 
+            this.Controls.Add(this.BUT_refreshTable);
             this.Controls.Add(this.chk_modified);
             this.Controls.Add(this.BUT_commitToFlash);
             this.Controls.Add(this.label2);
@@ -261,5 +270,6 @@
         private System.Windows.Forms.Label label2;
         private Controls.MyButton BUT_commitToFlash;
         private System.Windows.Forms.CheckBox chk_modified;
+        private Controls.MyButton BUT_refreshTable;
     }
 }