… GCSViews/ConfigurationView/ConfigPlanner.resx
… GCSViews/ConfigurationView/ConfigRawParams.resx
… GCSViews/ConfigurationView/ConfigRawParamsTree.cs
… GCSViews/ConfigurationView/ConfigRawParamsTree.resx