A new Planner setting checkbox called "Running on a slow computer" controls how the ConfigRawParams and ConfigRawParamsTree pages are reloaded. If enabled, it keeps the metadata info and refreshes only the values and the control. For safety, there is a button to refresh the whole table (without reloading parameters from the vehicle).

Time improvements: Measured on an Intel Baytrail T Z3735F Quad Core stick computer with 2GB RAM 
                                              original        Running on Slow Computer
ConfigRawParams reload        13sec                             2.4 sec
ConfigRawParamsTree reload  20sec                             3.5 sec

Note: Initial opening times of these pages are not improved; that needs a significant rewrite of the ParameterMetadataRpository, which is not in the scope yet.