@meee1 2021-12-20
this doesn't violate any cross thread gui calls does it?  ie the paralell.for's etc would