@EosBandi 2021-12-21
Paralell fors was there, I simply made the rowlist static. None of the two controls are used elsewhere and it just tests OK. So it did not break anything during testing.